% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_tests.R
\name{additional_tests}
\alias{additional_tests}
\title{Additional heteroscedasticity tests}
\description{
These functions extend the package with more specialized tests
for heteroscedasticity such as a studentized Breusch-Pagan test,
a bootstrap version of White's test and the Szroeter ordered test.
}
