% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performWhiteTest.R
\name{bootstrap_white_test}
\alias{bootstrap_white_test}
\title{Bootstrap implementation for White test}
\usage{
bootstrap_white_test(model, data, B, parallel, cross_products)
}
\description{
Bootstrap implementation for White test
}
\keyword{internal}
