\name{boston_housing}
\alias{boston_housing}
\docType{data}
\title{Boston housing prices dataset}
\description{Median housing values and associated predictors for Boston suburbs. This dataset is sourced from the \pkg{MASS} package.}
\usage{data(boston_housing, package = "heteroTests")}
\format{A data frame with 506 rows and 14 variables as described in \code{?MASS::Boston}.}
\source{\pkg{MASS} package.}
\examples{
data(boston_housing)
head(boston_housing)
}
