% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticsRegistry.R
\docType{data}
\name{.diagnostic_registry}
\alias{.diagnostic_registry}
\title{Diagnostics registry}
\format{
An object of class \code{environment} of length 16.
}
\usage{
.diagnostic_registry
}
\description{
Maintains a registry of diagnostic functions that can be called by
\code{runHeteroTests()} and \code{runDiagnostics()}. Users can register
custom diagnostics via \code{registerDiagnostic()}.
}
\keyword{datasets}
