% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performWhiteTest.R
\name{interpret_white_test}
\alias{interpret_white_test}
\title{Interpret White test results}
\usage{
interpret_white_test(result, alpha = 0.05)
}
\description{
Interpret White test results
}
\keyword{internal}
