% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_framework.R
\name{plot.power_analysis}
\alias{plot.power_analysis}
\title{Plot method for power analysis results}
\usage{
\method{plot}{power_analysis}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{simulate_power_analysis()}.}
}
\description{
Plot method for power analysis results
}
