% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostics.R
\name{plotDiagnosticSuiteEnhanced}
\alias{plotDiagnosticSuiteEnhanced}
\title{Enhanced diagnostic plot suite}
\usage{
plotDiagnosticSuiteEnhanced(model)
}
\arguments{
\item{model}{A fitted model of class \code{lm}.}
}
\value{
A list of ggplot objects.
}
\description{
Returns a list of improved diagnostic plots with statistical overlays.
}
