% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostics.R
\name{plotResidualDensity}
\alias{plotResidualDensity}
\title{Density plot of residuals}
\usage{
plotResidualDensity(model)
}
\arguments{
\item{model}{A fitted model of class \code{lm} or \code{glm}.}
}
\value{
A \code{ggplot} object.
}
\description{
Shows the distribution of residuals with a kernel density estimate.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + qsec, data = mtcars)
plotResidualDensity(m)
}
