% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostics.R
\name{plotResidualQQ}
\alias{plotResidualQQ}
\title{QQ plot of residuals}
\usage{
plotResidualQQ(model)
}
\arguments{
\item{model}{A fitted model of class \code{lm} or \code{glm}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualises departure from normality using a QQ plot.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + qsec, data = mtcars)
plotResidualQQ(m)
}
