% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostics.R
\name{plotResidualsFitted}
\alias{plotResidualsFitted}
\title{Plot residuals vs fitted values}
\usage{
plotResidualsFitted(model)
}
\arguments{
\item{model}{A fitted model of class \code{lm}.}
}
\value{
A \code{ggplot} object.
}
\description{
Generates a simple scatter plot of residuals against fitted values from a
linear model. A horizontal reference line at zero is added.
}
\details{
For a weighted fit the Pearson residuals \eqn{\sqrt{w_i}\, e_i} are plotted,
as in \code{plot.lm()}: those are the residuals that have constant variance when
the weights are right. The other residual plots in the package do the same.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + qsec, data = mtcars)
plotResidualsFitted(m)
}
