% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostics.R
\name{plotResidualsFittedEnhanced}
\alias{plotResidualsFittedEnhanced}
\title{Enhanced residuals vs fitted plot}
\usage{
plotResidualsFittedEnhanced(model)
}
\arguments{
\item{model}{A fitted model of class \code{lm}.}
}
\value{
A \code{ggplot} object.
}
\description{
Highlights influential observations and includes a LOESS smooth with
confidence bands.
}
