% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostics.R
\name{plotSpreadLevel}
\alias{plotSpreadLevel}
\title{Spread-Level plot for variance diagnostics}
\usage{
plotSpreadLevel(model)
}
\arguments{
\item{model}{A fitted model of class \code{lm}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots the square root of the absolute residuals against fitted values.
A lowess smooth is added to highlight trends.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + qsec, data = mtcars)
plotSpreadLevel(m)
}
