% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticsRegistry.R
\name{registerPlot}
\alias{registerPlot}
\title{Register a diagnostic plot}
\usage{
registerPlot(name, fun)
}
\arguments{
\item{name}{Name of the plot}

\item{fun}{Function taking a model and returning a ggplot object}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Register a diagnostic plot
}
\examples{
registerPlot("custom_plot", function(model) ggplot2::ggplot())
}
