% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticsRegistry.R
\name{runDiagnosticPlots}
\alias{runDiagnosticPlots}
\title{Run registered diagnostic plots}
\usage{
runDiagnosticPlots(
  model,
  plots = c("residuals_fitted", "spread_level", "density", "qq", "bubble_variance")
)
}
\arguments{
\item{model}{A fitted \code{lm} or \code{glm} object}

\item{plots}{Character vector of plot names to generate}
}
\value{
Named list of ggplot objects
}
\description{
Run registered diagnostic plots
}
\examples{
runDiagnosticPlots(lm(mpg ~ wt, mtcars))
}
