% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hetero.R
\name{sigma_exp_decay}
\alias{sigma_exp_decay}
\title{Exponential decay: sigma(x) = a * exp(-b * x)}
\usage{
sigma_exp_decay(x, a = 2, b = 0.2)
}
\arguments{
\item{x}{Numeric vector}

\item{a}{Scale factor}

\item{b}{Rate of decay}
}
\value{
Numeric vector
}
\description{
Exponential decay: sigma(x) = a * exp(-b * x)
}
