% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hetero.R
\name{sigma_inverse}
\alias{sigma_inverse}
\title{Inverse relationship: sigma(x) = a / (b + x)}
\usage{
sigma_inverse(x, a = 1, b = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{a}{Scale parameter}

\item{b}{Positive shift}
}
\value{
Numeric vector
}
\description{
Inverse relationship: sigma(x) = a / (b + x)
}
