% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hetero.R
\name{sigma_linear}
\alias{sigma_linear}
\title{Linear increase: sigma(x) = 0.5 + 0.2 * x}
\usage{
sigma_linear(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Linear increase: sigma(x) = 0.5 + 0.2 * x
}
