% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hetero.R
\name{sigma_logistic}
\alias{sigma_logistic}
\title{Logistic variance: sigma(x) = L / (1 + exp(-k * (x - x0)))}
\usage{
sigma_logistic(x, L = 2, k = 1, x0 = 5)
}
\arguments{
\item{x}{Numeric vector}

\item{L}{Upper asymptote}

\item{k}{Growth rate}

\item{x0}{Midpoint}
}
\value{
Numeric vector
}
\description{
Logistic variance: sigma(x) = L / (1 + exp(-k * (x - x0)))
}
