% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hetero.R
\name{sigma_multiplicative}
\alias{sigma_multiplicative}
\title{Multiplicative: variance proportional to |mu(x)|^p where mu(x)=beta0+beta1*x}
\usage{
sigma_multiplicative(x, mu_func, p = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{mu_func}{Function returning mean mu for given x}

\item{p}{Exponent controlling relationship}
}
\value{
Numeric vector
}
\description{
Multiplicative: variance proportional to |mu(x)|^p where mu(x)=beta0+beta1*x
}
