% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hetero.R
\name{sigma_piecewise_linear}
\alias{sigma_piecewise_linear}
\title{Piecewise linear variance with break at \code{thr}}
\usage{
sigma_piecewise_linear(x, thr = 5, slope1 = 0.1, slope2 = 0.3)
}
\arguments{
\item{x}{Numeric vector}

\item{thr}{Break point}

\item{slope1}{Slope below the break}

\item{slope2}{Slope above the break}
}
\value{
Numeric vector
}
\description{
Piecewise linear variance with break at \code{thr}
}
