% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hetero.R
\name{sigma_poly}
\alias{sigma_poly}
\title{Polynomial variance: sigma(x) = a + b * x + c * x^2}
\usage{
sigma_poly(x, a = 0.5, b = 0.1, c = 0.02)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Polynomial variance: sigma(x) = a + b * x + c * x^2
}
