% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hetero.R
\name{sigma_power}
\alias{sigma_power}
\title{Power variance: sigma(x) = a * |x|^p}
\usage{
sigma_power(x, a = 0.5, p = 0.5)
}
\arguments{
\item{x}{Numeric vector}

\item{a}{Scale factor}

\item{p}{Power exponent}
}
\value{
Numeric vector
}
\description{
Power variance: sigma(x) = a * |x|^p
}
