% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hetero.R
\name{sigma_sin}
\alias{sigma_sin}
\title{Sinusoidal variance: sigma(x) = A + B * sin(omega * x + phi)}
\usage{
sigma_sin(x, A = 1, B = 0.5, omega = 2 * pi/10, phi = 0)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Sinusoidal variance: sigma(x) = A + B * sin(omega * x + phi)
}
