% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hetero.R
\name{sigma_spatial}
\alias{sigma_spatial}
\title{Spatial heteroscedasticity: sigma(s) = gamma0 + gamma1 * distance from origin}
\usage{
sigma_spatial(coords, gamma0 = 0.5, gamma1 = 0.1)
}
\arguments{
\item{coords}{Matrix or data.frame with columns x and y}

\item{gamma0}{Numeric intercept}

\item{gamma1}{Numeric slope}
}
\value{
Numeric vector of sigma values
}
\description{
Spatial heteroscedasticity: sigma(s) = gamma0 + gamma1 * distance from origin
}
