% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hetero.R
\name{sigma_step}
\alias{sigma_step}
\title{Threshold step: low variance below \code{thr}, high variance above}
\usage{
sigma_step(x, thr = 5, low = 0.5, high = 2)
}
\arguments{
\item{x}{Numeric vector}

\item{thr}{Threshold value}

\item{low}{Numeric variance below \code{thr}}

\item{high}{Numeric variance above \code{thr}}
}
\value{
Numeric vector
}
\description{
Threshold step: low variance below \code{thr}, high variance above
}
