% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hetero.R
\name{sigma_u_shape}
\alias{sigma_u_shape}
\title{U-shaped variance: sigma(x) = a + b * (x - center)^2}
\usage{
sigma_u_shape(x, a = 0.1, b = 0.05, center = 5)
}
\arguments{
\item{x}{Numeric vector}

\item{a}{Base level}

\item{b}{Curvature}

\item{center}{Location of minimum variance}
}
\value{
Numeric vector
}
\description{
U-shaped variance: sigma(x) = a + b * (x - center)^2
}
