% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{std_warning}
\alias{std_warning}
\title{Generate standardized warning message}
\usage{
std_warning(type, ...)
}
\arguments{
\item{type}{Warning message type.}

\item{...}{Named arguments for the template.}
}
\value{
Issues a warning with a formatted message.
}
\description{
Generate standardized warning message
}
