\name{summary.HeteroDiagnostic}
\alias{summary.HeteroDiagnostic}
\title{Summarize diagnostics}
\usage{\method{summary}{HeteroDiagnostic}(object, ...)}
\arguments{\item{object}{A \code{HeteroDiagnostic} object.}}
\value{Named numeric vector of test statistics.}
\description{Summarize diagnostic test results.}
