\name{test.HeteroDiagnostic}
\alias{test}
\alias{test.HeteroDiagnostic}
\title{Run diagnostic tests}
\usage{test(object, ...)}
\arguments{
  \item{object}{A \code{HeteroDiagnostic} object.}
  \item{...}{Further arguments passed to the individual tests.}
}
\value{List of test results.}
\description{Run the registered diagnostics on the model.}
