% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_factory.R
\name{TestFactory}
\alias{TestFactory}
\title{Heteroscedasticity Test Factory}
\description{
Provides a registry of heteroscedasticity tests with metadata and
convenience methods to run them with input validation. This is an
alternative to the older environment-based registry used by
\code{runHeteroTests()}.
}
\section{Usage}{

The shared instance \code{test_factory} can be used to register new tests
and execute them. See \code{register} and \code{run_test} methods
for details.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TestFactory-register}{\code{TestFactory$register()}}
\item \href{#method-TestFactory-get_available}{\code{TestFactory$get_available()}}
\item \href{#method-TestFactory-run_test}{\code{TestFactory$run_test()}}
\item \href{#method-TestFactory-clone}{\code{TestFactory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFactory-register"></a>}}
\if{latex}{\out{\hypertarget{method-TestFactory-register}{}}}
\subsection{Method \code{register()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFactory$register(name, func, metadata = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the test}

\item{\code{func}}{Function with arguments \code{model} and \code{data}}

\item{\code{metadata}}{Optional list with metadata fields}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the factory
Get available tests
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFactory-get_available"></a>}}
\if{latex}{\out{\hypertarget{method-TestFactory-get_available}{}}}
\subsection{Method \code{get_available()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFactory$get_available(data_type = NULL, min_n = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_type}}{Filter by supported data type}

\item{\code{min_n}}{Filter by minimum observations}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character vector of test names
Run a registered test
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFactory-run_test"></a>}}
\if{latex}{\out{\hypertarget{method-TestFactory-run_test}{}}}
\subsection{Method \code{run_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFactory$run_test(test_name, model, data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test_name}}{Name of the test}

\item{\code{model}}{Fitted model}

\item{\code{data}}{Data frame used to fit the model}

\item{\code{...}}{Additional arguments passed to the test}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result of the test
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestFactory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TestFactory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestFactory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
