% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkModel}
\alias{checkModel}
\title{Validate model argument}
\usage{
checkModel(model)
}
\arguments{
\item{model}{An object to check.}
}
\value{
Invisible \code{model} if valid, otherwise an error is thrown.
}
\description{
Checks that \code{model} is of class \code{lm} or \code{glm}.
}
\keyword{internal}
