% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTeachingData.R
\name{downloadTeachingData}
\alias{downloadTeachingData}
\title{Download example teaching dataset}
\usage{
downloadTeachingData(name = "tips", destfile = tempfile(fileext = ".csv"))
}
\arguments{
\item{name}{Name of the dataset to download. Only "tips" is currently
supported.}

\item{destfile}{Optional path to save the downloaded dataset. Defaults to a
temporary file.}
}
\value{
Path to the downloaded dataset.
}
\description{
Provides convenient access to example datasets hosted online. Currently
supports the 'tips' dataset used in various tutorials.
}
\examples{
path <- downloadTeachingData("tips")
read.csv(path)
}
