% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_tests.R
\name{performStudentizedBPTest}
\alias{performStudentizedBPTest}
\title{Studentized Breusch-Pagan test}
\usage{
performStudentizedBPTest(model, data)
}
\arguments{
\item{model}{A fitted model of class \code{lm}.}

\item{data}{The data frame used to fit \code{model}.}
}
\value{
An object of class \code{htest} with the test statistic and p-value.
}
\description{
Performs the Breusch-Pagan test using studentized residuals,
providing robustness to non-normality.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + qsec, data = mtcars)
performStudentizedBPTest(m, mtcars)
}
