% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_tests.R
\name{performSzroeterTest}
\alias{performSzroeterTest}
\title{Szroeter test for ordered alternatives}
\usage{
performSzroeterTest(model, data, order_by)
}
\arguments{
\item{model}{A fitted model of class \code{lm}.}

\item{data}{The data frame used to fit \code{model}.}

\item{order_by}{Variable name to order the observations by.}
}
\value{
An object of class \code{htest}.
}
\description{
Detects monotonic changes in variance when the observations can be
ordered by a known variable.
}
