% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_tests.R
\name{performWhiteTestBootstrap}
\alias{performWhiteTestBootstrap}
\title{Bootstrap White test}
\usage{
performWhiteTestBootstrap(model, data, B = 1000, parallel = FALSE)
}
\arguments{
\item{model}{A fitted model of class \code{lm}.}

\item{data}{The data frame used to fit \code{model}.}

\item{B}{Number of bootstrap replications.}

\item{parallel}{Logical, run in parallel using the \code{parallel} package?}
}
\value{
An object of class \code{htest} with the bootstrap p-value.
}
\description{
Uses bootstrap resampling of residuals to estimate the distribution
of White's test statistic. Useful for small samples.
}
