% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostics.R
\name{plotResidualsFitted}
\alias{plotResidualsFitted}
\title{Plot residuals vs fitted values}
\usage{
plotResidualsFitted(model)
}
\arguments{
\item{model}{A fitted model of class \code{lm}.}
}
\value{
A \code{ggplot} object.
}
\description{
Generates a simple scatter plot of residuals against fitted values from a
linear model. A horizontal reference line at zero is added.
}
\examples{
data(mtcars)
m <- lm(mpg ~ wt + qsec, data = mtcars)
plotResidualsFitted(m)
}
