% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performWhiteTest.R
\name{print.htest_enhanced}
\alias{print.htest_enhanced}
\title{Print method for enhanced White test results}
\usage{
\method{print}{htest_enhanced}(x, ...)
}
\arguments{
  \item{x}{An object of class \code{htest_enhanced}, as returned by
    \code{performWhiteTestEnhanced()}.}
  \item{...}{Further arguments passed to \code{print.htest()}.}
}
\description{
Print method for enhanced White test results
}
