% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{std_error}
\alias{std_error}
\title{Generate standardized error message}
\usage{
std_error(type, ...)
}
\arguments{
\item{type}{Error message type.}

\item{...}{Named arguments to replace in the template.}
}
\value{
Stops execution with a formatted message.
}
\description{
Generate standardized error message
}
\keyword{internal}
